refactor: modularize stats update functions
- Create src/lib/stats directory for individual stat update functions
- Implement updateGameHours and updateRepoStars as separate modules
- Refactor update-stats cron job to use new modular structure
- Improve error handling and logging in stat update functions
