+++ src/lib/stats/updateGameHours.ts
@@ -1,4 +1,3 @@
-import { NextResponse, type NextRequest } from "next/server";
 import { kv } from "@vercel/kv";
 
 import { steamConfig } from "@/lib/constants";
@@ -7,30 +6,20 @@ const STEAM_API_KEY = process.env.STEAM_API_KEY;
 const STEAM_ID = steamConfig.STEAM_ID;
 const RUST_APP_ID = 252490;
 
-export async function GET(req: NextRequest) {
-  const authHeader = req.headers.get("authorization");
-  if (authHeader !== `Bearer ${process.env.CRON_SECRET}`) {
-    return new Response("Unauthorized", {
-      status: 401,
-    });
-  }
+export async function updateGameHours() {
   const url = `http://api.steampowered.com/IPlayerService/GetOwnedGames/v0001/?key=${STEAM_API_KEY}&steamid=${STEAM_ID}&format=json`;
 
   try {
-    const res = await fetch(url, {
-      cache: "no-store",
-    });
+    const res = await fetch(url, { cache: "no-store" });
     const data = await res.json();
     const gameHours = Math.floor(
       data.response.games.find((game: any) => game.appid === RUST_APP_ID)
         .playtime_forever / 60,
     );
     await kv.set("gameHours", gameHours);
-    return NextResponse.json({ gameHours }, { status: 200 });
+    return gameHours;
   } catch (error) {
-    return NextResponse.json(
-      { error: "Data currently unavailable" },
-      { status: 500 },
-    );
+    console.error("Error updating game hours:", error);
+    throw error;
   }
 }
